requirements.txt
Kivy==2.2.1
kivymd==1.1.1
requests==2.31.0
pillow==10.1.0
python-dotenv==1.0.0
cryptography==41.0.7
bcrypt==4.1.2
buildozer==1.5.0
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyJWT==2.8.0
gunicorn==21.2.0
